Fixed Dirb Installer - needed to modify framework.py and switch the order of two lines in order for this fix to work. Fix sugguested by oliverloveless "The issue is that ptf makes assumptions on whether the repo needs to be compiled before a launcher can be created.
ptf assumes that all git repos are not compiled, and therefore thinks it's ok to create the launcher, before the after_commands (as specified in the config file) are run. This causes an issue in any compiled tool originating from github, because they don't have the appropriately named executable file in the repository (which src.core.launcher() uses in the launcher creation process)."

Fixed Fluxion installer - deltaflux took his repo offline due to laws in his country but put a copy up on his website which this installer now grabs.

Permissons on msfdb.sh automatically changed during program execution - I did not manually  chmod this -  it happened after running the installer - but the updated permissions are carrying over to this commit